+++ GameEngine/GameEngine/Engine/Render/ProgramResources/Types/BlockResources.cpp
@@ -6,7 +6,7 @@
 
 namespace AGE
 {
-	BlockResources::BlockResources(GLint id, std::string &&name, GLenum type, glm::vec3 const &info) :
+	BlockResources::BlockResources(GLint id, std::string &&name, GLenum type, glm::uvec3 const &info) :
 		AProgramResources(id, std::move(name), type),
 		_parent(nullptr),
 		_offset(info.x),
